Refactored locus and population modules into the panel module
